downstream/requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
attr==0.3.2
attrs==23.2.0
beartype==0.18.5
beautifulsoup4==4.12.3
blessed==1.20.0
boto3==1.34.97
botocore==1.34.97
causal-conv1d==1.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
connected-components-3d==3.14.1
contourpy==1.2.1
croniter==1.3.15
cycler==0.12.1
dateutils==0.6.12
deepdiff==7.0.1
docker-pycreds==0.4.0
editor==1.6.6
einops==0.8.0
ema-pytorch==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.88.0
fastremap==1.14.1
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.48.2
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.23.0
idna==3.4
imagenetv2-pytorch==0.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inquirer==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
kiwisolver==1.4.5
lightning==1.9.5
lightning-cloud==0.5.68
lightning-utilities==0.11.2
littleutils==0.2.4
mamba-ssm==1.0.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
monai==1.3.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nibabel==5.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.3
ogb==1.3.6
opencv-python==4.10.0.84
openpyxl==3.1.2
ordered-set==4.1.0
outdated==0.2.2
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pip==23.3.1
protobuf==3.20.3
psutil==5.9.8
pydantic==1.10.15
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==2.1.2
pytz==2024.1
PyYAML==6.0.1
readchar==4.0.6
regex==2024.5.15
requests==2.28.1
rich==13.7.1
runs==1.2.2
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.0.1
setproctitle==1.3.3
setuptools==68.2.2
SimpleITK==2.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
timm==0.9.16
tokenizers==0.13.3
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchcam==0.4.0
torchmetrics==1.3.2
torchvision==0.16.0+cu121
tqdm==4.66.4
traitlets==5.14.3
transformers==4.30.1
triton==2.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
uvicorn==0.29.0
vector-quantize-pytorch==1.1.2
wandb==0.16.6
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.1.2
wheel==0.41.2
wilds==2.0.0
XlsxWriter==3.2.0
xmod==1.8.1
yarl==1.9.4
zipp==3.17.0